Update subscribers table.
This SQL is a bit more complicated for a few reasons:

1) We have to do some parsing on the date value. ISO8601 is notoriously batshit
to parse using datetime.strptime; this dateutil library is what I've most often
used to work with these dates in production. It's very magical but ISO8601 is
sufficiently precise that I wouldn't expect the library to get "confused" by,
say, ambiguous dates, like 01-01-01.

2) We have two "modes": canceled and new. In the first case, we do an INSERT; in
the second, an UPDATE. We do in fact allow this query to fail, even though it is
plausible that a user may get his/her row "updated" without an attendant
cancellation.  (For example, suppose a user is on the receiving end of a
customer loyalty discount. This would probably arrive at our system as a 'new'
order with a different price from the existing price. As written, our code would
reject it.) I consider these sorts of edge cases outside the scope of this
exercise.