Do primitive error handling.
If the user uploads a canceled purchase which we don't recognize, raise an
UnknownSubscripionError; this is one of the "stretch goals". If we get an error
at any stage of the databse transcation, rollback and proceed. Return all of the
"failed" rows in a JSON document. I've done my best to capture any kind of
400-class error, e.g. missing fields.